requirements.txt
crowdstrike-falconpy
aniso8601
pandas
openpyxl